+++ packages/flame/CHANGELOG.md
@@ -9,6 +9,10 @@ Refer to the [CONTRIBUTING guide](https://github.com/lightspeed/flame/blob/maste
 
 ## [Unreleased]
 
+## Breaking
+
+- Remove SVGs from the mainline repo. Split of icons/flags into their own package.
+
 ### Fixed
 
 - Typecast Button to bypass weird automatic type generation shenanigans ([#125](https://github.com/lightspeed/flame/pull/125))